Move session loading into separate module
