+++ index.html
@@ -1041,6 +1041,7 @@ chrom1,start1,stop1,chrom2,start2,stop2,variant_name,score,strand1,strand2,varia
 <!-- Index -->
 <script type="module" src="js/index_splitthreader.js"></script>
 <script type="module" src="js/index_ribbon.js"></script>
+<script type="module" src="js/index.js"></script>
 
 </body>
 </html>
